Refactor: create simu package, remove useless simu dependencie in car
